pyproject.toml
[project]
name = "a_share_rag_project"
version = "0.1.0"
description = "A股上市公司关联性RAG项目"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-igraph>=0.11.0",
    "lancedb>=0.24.0",
    "transformers>=4.42.0",
    "sentence-transformers>=3.0.1",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.30.0",
    "openai>=1.0.0",
    "pyvis>=0.3.0",
    "rank-bm25>=0.2.0",
    "httpx>=0.28.1",
    "filelock>=3.18.0",
    "tenacity>=9.1.2",
    "torch",
    "psutil",
    "pandas",
    "pyarrow",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "ruff>=0.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "DTZ",  # flake8-datetimez
    "T10",  # flake8-debugger
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.2",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"